Organising and neatening SDK layout, fixing naming
