+++ Assets/Plugins/EuclideonUdSDK/udSDK/csAPI/udAttributes.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 50f061076f4404b49931849e99c94aee
+guid: 8a14867eaa8670f41b21a3ba317bfc0b
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
