+++ Assets/Plugins/EuclideonUdSDK/udSDK/csAPI/udContext.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1412ac0ee01136947859c499fb203a98
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
