+++ Assets/Plugins/EuclideonUdSDK/udSDK/csAPI/udConvert.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3fd4b80e0f08c7642853018b780afca9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
