+++ Assets/Plugins/EuclideonUdSDK/udSDK/csAPI/udError.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 06ff0a392c5bc6a4eb6f790ee3e83c44
+guid: 182481adf60d60145b3ef980e63aab5a
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
