+++ Assets/Plugins/EuclideonUdSDK/udSDK/csAPI/udPointCloud.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: bb00b273f073254449b732eb414fa243
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
