+++ Assets/Plugins/EuclideonUdSDK/udSDK/csAPI/udProject.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 46c639d22201b0540a629e7d3bcf60bd
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
