+++ Assets/Plugins/EuclideonUdSDK/udSDK/csAPI/udQueryContext.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 73ed5dbb728ec9047ad1588ed2d9eb80
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
