+++ Assets/Plugins/EuclideonUdSDK/udSDK/csAPI/udRenderContext.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b71a86efc053b5044ba209718cf696b3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
