+++ Assets/Plugins/EuclideonUdSDK/udSDK/csAPI/udRenderTarget.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 252ef56cb0f29da438a50f60b93f10a2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
