+++ Assets/Plugins/EuclideonUdSDK/udSDK/csAPI/udVersion.cs
@@ -1,0 +1,8 @@
+using System;
+using System.Runtime.InteropServices;
+using UnityEngine;
+
+namespace udSDK
+{
+    
+}
