+++ Assets/Plugins/EuclideonUdSDK/udSDK/csAPI/udVersion.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: dccd9c345bcc3514da9069d3bc6dcc3b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
